Relocate handlers.ts and node.ts files
