+++ src/entities/movies/hooks/useMovies.test.ts
@@ -3,8 +3,8 @@ import { moviesMock } from "../mocks/moviesMock";
 import { MovieStructure } from "../types";
 import useMovies from "./useMovies";
 import { wrapWithProviders } from "../../../utils/testUtils";
-import { server } from "../../../mocks/moviesMocks/node";
-import { errorHandlers } from "../../../mocks/moviesMocks/handlers";
+import { server } from "../../../mocks/node";
+import { errorHandlers } from "../../../mocks/handlers";
 
 describe("Given a getMovies function", () => {
   describe("When it is invoked", () => {
